refactor: extract GroupMemberFields of GroupForm
